requirements.txt
langchain-community       # for WebBaseLoader, FAISS wrapper
langchain-core                 # for prompts, chains                   # BeautifulSoup parsing
faiss-cpu                   # vector index backend
openai
langchain
langgraph
chromadb
docker
tqdm
langchain_openai
fastapi
uvicorn
python-dotenv